pyproject.toml
[project]
name = "algorithms"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pillow>=11.2.1",
    "scikit-image>=0.25.2",
    "torch==2.6.0",
    "torchvision==0.21.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://mirror.sjtu.edu.cn/pytorch-wheels/cpu"
explicit = true

[[tool.uv.index]]
name = "mirror.sjtu.edu.cn"
url = "https://mirror.sjtu.edu.cn/pypi/web/simple"
default = true


[tool.uv.sources]
torch = [
  { index = "pytorch-cpu" },
]
torchvision = [
  { index = "pytorch-cpu" },
]
